requirements.txt
aiogram==3.7.0
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.30
